using handelbars template language and express to generate web view for / , refactored search to api
