+++ views/index.hbs
@@ -1,0 +1,13 @@
+{{> welcomemenu }}
+
+<section class="ui stacked segment">
+    <div class="ui grid">
+        <aside class="six wide column">
+            <img src="images/kujira.jpg" class="ui medium image">
+        </aside>
+        <article class="ten wide column">
+            <header class="ui  header">Kujira</header>
+            <p> You can login here to use Kujira! </p>
+        </article>
+    </div>
+</section>
