+++ views/partials/welcomemenu.hbs
@@ -1,0 +1,7 @@
+<nav class="ui inverted blue menu">
+    <header class="header item"> <a href="/"> Kujira </a> </header>
+    <div class="right menu">
+        <!--<a class="item" href="/signup"> Signup</a>-->
+        <a class="item" href="/login">  Login</a>
+    </div>
+</nav>
